🐛fix(verification): PR #97 리뷰 반영 (data.go.kr serviceKey 이중 인코딩 외 6건)
RestClient uri 템플릿이 serviceKey를 재인코딩하므로 URLEncoder 사전 인코딩 제거(이중 인코딩 시 키 거부). prefilter 0점 후보 제외, FidelityItem 분리(미사용 래퍼 제거), stance trim+Locale.ROOT 정규화, retrieve topK는 URL 검증 후 적용(사전 truncate 제거), VerifySourceConverter Locale.ROOT, EvidenceDto는 응답 DTO 규약대로 Lombok class화.
